:recycle: refactor(domain): 타입 정의를 위해 interface 사용
